backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "3.11.3"
uvicorn = "^0.22.0"
fastapi = "^0.100.0"
httpx = "^0.24.1"
pydantic = {extras = ["email"], version = "^2.1.1"}
gunicorn = "^20.1.0"
pytest = "^7.4.0"
elasticsearch = "8.8.0"
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pytest-asyncio = "^0.21.0"
freezegun = "^1.2.2"
sqlalchemy = "^2.0.19"
psycopg2 = "^2.9.6"
pydantic-settings = "^2.0.2"
pygit2 = "^1.12.2"
celery = "^5.3.1"
redis = "^4.6.0"
pygithub = "^1.59.0"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
isort = "^5.12.0"
autoflake = "^2.1.1"
flake8 = "^6.0.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pre-commit = "^3.3.3"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "users_file: provide users data to the users_file fixture"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.masonry.api"
